% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{CellResponseTypeAvailabilityCheck}
\alias{CellResponseTypeAvailabilityCheck}
\title{Checking CellResponseType Availability}
\usage{
CellResponseTypeAvailabilityCheck(OBJ, RESP)
}
\arguments{
\item{OBJ}{A ForeseeTrain object that you want to check the CellResponseType availability in.}

\item{RESP}{CellResponseType to be checked.}
}
\value{
\item{Returns an invisible TRUE if RESP is available in OBJ, if not available an error will be generated}{}
}
\description{
Checking If CellResponseType is Available in Object
}
\examples{
CellResponseTypeAvailabilityCheck(DAEMEN,"GI50")
}
