% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplicationHandler.R
\name{DuplicationHandler}
\alias{DuplicationHandler}
\title{Remove Duplicated Gene Names from a FORESEE Object}
\usage{
DuplicationHandler(Object, DuplicationHandling)
}
\arguments{
\item{Object}{FORESEE Object (ForeseeCell or ForeseeTrain) that contains all data needed to train a model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{DuplicationHandling}{Method for handling duplicates of gene names, such as taking none, the average, the first hit, etc.}
}
\value{
\item{Object}{The object without duplicated gene names}
}
\description{
DuplicationHandler removes duplicates in the gene names from the FORESEE Object.
}
