% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCombiner.R
\name{FeatureCombiner}
\alias{FeatureCombiner}
\title{FeatureCombiner}
\usage{
FeatureCombiner(TrainObject, TestObject, InputDataTypes)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{TestObject}{Object that contains all data that the model is to be tested on, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{InputDataTypes}{Data types of the TrainObject that are to be used to train the model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}
}
\value{
\item{TrainObject}{The TrainObject with a new Feature matrix combining all specified input data types and a Featuretype Vector indicating the molecular data type of each feature}
        \item{TestObject}{The TestObject with a new Feature matrix combining all specified input data types and a Featuretype Vector indicating the molecular data type of each feature}
}
\description{
The Feature Combiner combines all selected Input Data Types into one Feature Matrix
}
\examples{
FeatureCombiner(GDSC,GSE6434,c("GeneExpression", "Mutation"))
}
