% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelector.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{Select the Genes that are used as Model Features}
\usage{
FeatureSelector(TrainObject, DrugName, GeneFilter)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{DrugName}{Name of the drug whose efficacy is supposed to be predicted with the model}

\item{GeneFilter}{Set of genes to be considered for training the model, such as all, a certain percantage based on variance or p-value, specific gene sets like landmark genes, gene ontologies or pathways, etc.}
}
\value{
\item{TrainObject}{The TrainObject with the selected gene set as features.}
}
\description{
The FeatureSelector selects features from all genes to be used for drug efficacy prediction.
}
