% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foreseer.R
\name{Foreseer}
\alias{Foreseer}
\title{Apply the trained ForeseeModel on a TestObject}
\usage{
Foreseer(TestObject, ForeseeModel, BlackBox)
}
\arguments{
\item{TestObject}{Object that contains all data that the model is to be tested on, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{ForeseeModel}{Model that has been trained on a TrainObject with the function ForeseeTrain.}
}
\value{
\item{Foreseen}{Predicted drug response of the samples listed in the TestObject obtained by applying the ForeseeModel to the molecular data.}
}
\description{
The Foreseer applies the ForeseeModel that was trained on a FORESEE TrainObject to the test data to gain a prediction of the TestObject's response.
}
