% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDSC}
\alias{GDSC}
\title{Genomics of Drug Sensitivity in Cancer or GDSC}
\format{ForeseeCell object
\describe{
  \item{GeneExpression}{Gene Expression Matrix}
  \item{AUC}{...}
  ...
}}
\source{
\url{http://www.cancerrxgene.org}
}
\usage{
GDSC
}
\description{
A dataset containing drug responses on cancer celllines.
}
\keyword{datasets}
