% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellResponseData.R
\name{GetCellResponseData}
\alias{GetCellResponseData}
\title{Get Cell Line Drug Response from FORESEE Object}
\usage{
GetCellResponseData(TrainObject, DrugName, CellResponseType)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{CellResponseType}{Format of the drug response data of the TrainObject, such as IC50, AUC, GI50, etc., that is used for prediction}
}
\value{
\item{TrainObject}{The TrainObject with extracted drug response data.}
}
\description{
The GetCellResponseData Function extracts the features of the drug response data that are relevant to predict the drug response.
}
\examples{
GetCellResponseData(GDSC,"Gemcitabine","AUC")
}
