% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HK_genes_entrez}
\alias{HK_genes_entrez}
\title{Human Housekeeping Genes}
\format{integers}
\source{
\url{https://www.tau.ac.il/~elieis/HKG/}
}
\usage{
HK_genes_entrez
}
\description{
A list of genes (ENTREZ IDs) that are involved in basic cell maintenance and, therefore, are expected to maintain constant expression levels in all cells and conditions.
Published by E. Eisenberg and E.Y. Levanon, Trends in Genetics, 29 (2013) "Human housekeeping genes revisited"
}
\keyword{datasets}
