% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homogenizer.R
\name{Homogenizer}
\alias{Homogenizer}
\title{Homogenize a Pair of two FORESEE Objects}
\usage{
Homogenizer(TrainObject, TestObject, HomogenizationMethod)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{TestObject}{Object that contains all data that the model is to be tested on, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{HomogenizationMethod}{Method for homogenizing data of the TrainObject and TestObject, such as ComBat, quantile normalization, limma, RUV, etc.}
}
\value{
\item{TrainObject}{The TrainObject with homogenized features according to the chosen TestObject.}
        \item{TestObject}{The TestObject with homogenized features according to the chosen TrainObject.}
}
\description{
The Homogenizer function removes batch effects and homogenizes the data of a given TrainObject and TestObject.
}
