% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{listCellLines}
\alias{listCellLines}
\title{List All Cell lines Inside a ForeseeTrain Object}
\usage{
listCellLines(OBJ)
}
\arguments{
\item{OBJ}{A ForeseeTrain object of which you want to extract its containing cell lines.}
}
\value{
\item{Character vector of all cell line names in OBJ.}{}
}
\description{
listCellLines returns all cell lines available in a ForeseeTrain Object.
}
\examples{
listCellLines(GDSC)
}
