% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{listDrugs}
\alias{listDrugs}
\title{List All Drugs Inside a ForeseeTrain Object}
\usage{
listDrugs(OBJ)
}
\arguments{
\item{OBJ}{A ForeseeTrain object that you want to extract all possible drugs it has response information on.}
}
\value{
\item{Character vector of all drug names in OBJ.}
}
\description{
listDrugs returns all possible drugs that were tested and has reponses available in a ForeseeTrain Object.
}
\examples{
listDrugs(GDSC)
}
