% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{listInputOptions}
\alias{listInputOptions}
\title{List Input Options for FORESEE Methods}
\usage{
listInputOptions(FunArgument, OBJ)
}
\arguments{
\item{FunArgument}{Character string of the input argument (for DrugName, CellResponseType, InputDataTypes, TrainingTissue
and TestingTissue) or the function corresponding to the input of ForeseeTrain (CellResponseTransformation, DuplicationHandling,
HomogenizationMethod, GeneFilter, FeaturePreprocessing and BlackBox) or ForeseeTest (Evaluation).
Check the help of ForeseeTrain and ForeseeTest for more information.}

\item{OBJ}{A ForeseeTrain or ForeseeTest object that you want to extract options for. Only necessary for when FunArgument is
DrugName, CellResponseType, InputDataTypes, TrainingTissue or TestingTissue.}
}
\value{
\item{Character vector of all possible inputs}{}
}
\description{
listInputOptions returns possible input arguments in ForeseeTrain (for DrugName, CellResponseType, InputDataTypes, TrainingTissue,
TestingTissue, CellResponseTransformation, DuplicationHandling, HomogenizationMethod, GeneFilter, FeaturePreprocessing
and BlackBox arguments) and in ForeseeTest (for the Evaluation argument).
}
\examples{
listInputOptions("CellResponseProcessor")
listInputOptions("DuplicationHandler")
listInputOptions("Homogenizer")
listInputOptions("Validator")
listInputOptions("DrugName", CCLE)[1:10]
listInputOptions("InputDataTypes", GAO)
listInputOptions("CellResponseType", GDSC)
listInputOptions("TrainingTissue", CCLE)
listInputOptions("TestingTissue", GSE33072_erlotinib)
}
