% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM_genes_entrez}
\alias{LM_genes_entrez}
\title{Landmark Genes of the Broad Institute (LINCS)}
\format{integers}
\source{
\url{https://clue.io/command?q=/gene-space%20lm
}
}
\usage{
LM_genes_entrez
}
\description{
A list of a subset of informative genes (ENTREZ IDs) determined by the Broad Institute.
Landmark genes were selected as those widely expressed across lineage and were found to have good predictive power for inferring the expression of other genes that are not directly measured in the L1000 assay.
}
\keyword{datasets}
