% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{requireForesee}
\alias{requireForesee}
\title{Loading/Attaching (and Installing) a Package}
\usage{
requireForesee(package)
}
\arguments{
\item{package}{the name of a package to be loaded and attached (and installed).}
}
\value{
requireForesee returns (invisibly) a logical indicating whether the required package was available (before installation attempts).
}
\description{
requireForesee is the same as 'require' from the base package, except in case of a missing package, it
tries to install it via 'biocLite' from Bioconductor. For installation to work, R needs to have access
to the internet (more precisely "https://bioconductor.org/biocLite.R" should be accessable to R).
}
\examples{
requireForesee(ranger)
}
