% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{Validator}
\alias{Validator}
\title{Validate the Performance of a ForeseeModel on a new TestObject}
\usage{
Validator(Foreseen, TestObject, Evaluation)
}
\arguments{
\item{Foreseen}{Predicted drug response of the TestObject obtained by applying the ForeseeModel.}

\item{TestObject}{Object that contains all data that the model is to be tested on, especially the true, measured drug response.}

\item{Evaluation}{Measure for evaluating the model performance, such as ROC-Curve, AUC or p-value of ROC-Curve, Rsquared, MSE, Correlation, F-Test, etc.}
}
\value{
\item{Performance}{Evaluation Measure of the Predictability of the ForeseeModel trained on the TrainObject and tested on the TestObject.}
}
\description{
The Validator evaluates the performance of the model by comparing the predicted response Foreseen to the reported, true response from the TestObject's Annotation.
}
