% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{CellorPatient}
\alias{CellorPatient}
\title{Test if the Object is a Cell or a Patient Based on the Drug Response Type}
\usage{
CellorPatient(Object)
}
\arguments{
\item{Object}{FORESEE TrainObject or TestObject that should be analyzed.}
}
\value{
\item{CellorPatient_status}{The Status of the FORESEE Object that states, whether it is a cell (true) or a patient (false).}
}
\description{
The CellorPatient Function tests if the object is a cell or a patient based on the type of the drug response annotation.
}
\examples{
CellorPatient(GDSC)
}
