% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePreprocessor.R
\name{FeaturePreprocessor}
\alias{FeaturePreprocessor}
\title{Preprocess the Inputs of both TrainObject and TestObject for Modeling Drug Response}
\usage{
FeaturePreprocessor(TrainObject, TestObject, FeaturePreprocessing)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{TestObject}{Object that contains all data that the model is to be tested on, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{FeaturePreprocessing}{Method for preprocessing the inputs of the model, such as z-score, principal component analysis, PhysioSpace similarity, etc.}
}
\value{
\item{TrainObject}{The TrainObject with preprocessed features.}
        \item{TestObject}{The TestObject with preprocessed features.}
}
\description{
The FeaturePreprocessor converts the original features into predictive features that are defined by FeaturePreprocessing.
}
