% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeaturePreprocessor.R
\name{FeaturePreprocessor}
\alias{FeaturePreprocessor}
\title{Preprocess the Gene Expression Inputs of both TrainObject and TestObject for Modeling Drug Response}
\usage{
FeaturePreprocessor(TrainObject, TestObject, FeaturePreprocessing)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type, etc. ) and drug response data}

\item{TestObject}{Object that contains all data that the model is to be tested on, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type, etc. ) and drug response data}

\item{FeaturePreprocessing}{Method for preprocessing the inputs of the model:
The function 'zscore_genewise' calculates the zscore normalizing each gene over all samples,
The function 'zscore_samplewise' calculates the zscore normalizing each sample over all genes,
The function 'pca' does principal component analysis,
The function 'physio' does physiospace analysis with the samples using cell line gene expression of the gdsc data base as physiological references,
The function 'none' keeps the gene expression values unchanged,
The function 'listInputOptions("FeaturePreprocessor")' returns a list of the possible options.
Instead of chosing one of the implemented options, a user-defined function can be used as an input.}
}
\value{
\item{TrainObject}{The TrainObject with preprocessed features.}
        \item{TestObject}{The TestObject with preprocessed features.}
}
\description{
The FeaturePreprocessor converts the original gene expression features into predictive features with a function defined by FeaturePreprocessing.
}
\examples{
FeaturePreprocessor(GDSC,GSE6434,"zscore_genewise")
}
