% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelector.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{Select the Genes that are used as Model Features}
\usage{
FeatureSelector(TrainObject, TestObject, GeneFilter)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{TestObject}{Object that contains all data that the model is to be tested on, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{GeneFilter}{Set of genes to be considered for training the model, such as all, a certain percantage based on variance or p-value, specific gene sets like landmark genes, gene ontologies or pathways, etc.
The option 'variance' removes the 20 % genes of lowest variance across samples in the TrainObject
The option 'pvalue' removes the 20 % genes of lowest p-value (ttest) across samples in the TrainObject
The option 'landmarkgenes' uses the L1000 gene set downloaded from CLUE Command App
The option 'ontology' uses a specific set of genes included in the chosen ontology? -> Ali
The option 'pathway' uses a specific set of genes included in the chosen pathway? -> Ali
The option 'all' keeps all genes as features
If the user inserts a list as an input, the list is considered as chosen features.}
}
\value{
\item{TrainObject}{The TrainObject with the selected gene set as features.}
        \item{TestObject}{The TestObject with homogenized features according to the chosen TrainObject.}
}
\description{
The FeatureSelector selects features from all genes to be used for drug efficacy prediction.
}
