% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foreseer.R
\name{Foreseer}
\alias{Foreseer}
\title{Apply the trained ForeseeModel on a TestObject}
\usage{
Foreseer(TestObject, ForeseeModel)
}
\arguments{
\item{TestObject}{Object that contains all data that the model is to be tested on, such as gene expression, mutation, copy number variation, methylation, cancer type, drug response data, etc.}

\item{ForeseeModel}{Model that has been trained on a TrainObject with ForeseeTrain.}
}
\value{
\item{Foreseen}{Predicted drug response of the TestObject obtained by applying the ForeseeModel.}
}
\description{
The Foreseer applies the ForeseeModel that was trained on a FORESEE TrainObject to the test data to gain a prediction of the TestObject's response.
}
