% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homogenizer.R
\name{Homogenizer}
\alias{Homogenizer}
\title{Homogenize a Pair of two FORESEE Objects}
\usage{
Homogenizer(TrainObject, TestObject, HomogenizationMethod)
}
\arguments{
\item{TrainObject}{Object that contains all data needed to train a model, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{TestObject}{Object that contains all data that the model is to be tested on, including molecular data (such as gene expression, mutation, copy number variation, methylation, cancer type) and drug response data}

\item{HomogenizationMethod}{Method for homogenizing data of the TrainObject and TestObject.
The function 'ComBat' uses the batch effect removal ComBat of the sva package,
The function 'quantile' uses the quantile normalization of the preprocessCore package,
The function 'limma' uses the removeBatchEffect function of the limma package,
The function 'YuGene' uses the function of the YuGene package by Le Cao,
The function 'RUV' regresses out unwanted variation using the 10 principal components of negative control genes (here: list of human housekeeping by Eisenberg and Levanon (2013))
The function 'none' doesn't do any batch effect correction.
The function 'listInputOptions("Homogenizer")' returns a list of the possible options.
Instead of choosing one of the implemented options, a user-defined function can be used as an input.}
}
\value{
\item{TrainObject}{The TrainObject with homogenized features according to the chosen TestObject.}
        \item{TestObject}{The TestObject with homogenized features according to the chosen TrainObject.}
}
\description{
The Homogenizer function removes batch effects and homogenizes the data of a given TrainObject and TestObject.
}
\examples{
Homogenizer(GDSC,GSE6434,"quantile")
}
