% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{listCellLines}
\alias{listCellLines}
\title{List All Cell lines Inside a ForeseeTrain Object}
\usage{
listCellLines(OBJ)
}
\arguments{
\item{OBJ}{A ForeseeTrain object of which you want to extract its containing cell lines.}
}
\value{
\item{Character vector of all cell line names in OBJ.}{}
}
\description{
listCellLines returns all cell lines (or sample names in case of a xenograft data set)
available in a ForeseeTrain Object (All cell lines that are included in gene expression matrix
of ForeseeTrain Object).
}
\examples{
listCellLines(GDSC)
listCellLines(WITKIEWICZ)
}
