% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{listInputOptions}
\alias{listInputOptions}
\title{List Input Options for FORESEE Methods}
\usage{
listInputOptions(METHOD)
}
\arguments{
\item{METHOD}{Character string of the function corrisponding to the input of ForeseeTrain or ForeseeTest.
Check the help of Foreseetrain and ForeseeTest for more info.}
}
\value{
\item{Character vector of all possible inputs}
}
\description{
listInputOptions returns possible input values in ForeseeTrain (CellResponseTransformation, DuplicationHandling,
HomogenizationMethod, GeneFilter, FeaturePreprocessing and BlackBox) and in ForeseeTest (Evaluation).
}
\examples{
listInputOptions("CellResponseProcessor")
listInputOptions("DuplicationHandler")
listInputOptions("Homogenizer")
listInputOptions("Validator")
}
